LPS-21299 Move User Groups as another Portlet (User Groups Admin) & LPS-20774 Can't create a user, organization, location or user group on View Users page of role
