+++ portal-impl/src/com/liferay/portlet/usergroupsadmin/action/ActionUtil.java
@@ -1,0 +1,55 @@
+/**
+ * Copyright (c) 2000-2011 Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portlet.usergroupsadmin.action;
+
+import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.model.UserGroup;
+import com.liferay.portal.service.UserGroupServiceUtil;
+import com.liferay.portal.util.PortalUtil;
+import com.liferay.portal.util.WebKeys;
+
+import javax.portlet.PortletRequest;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * @author Brian Wing Shun Chan
+ * @author Alexander Chow
+ */
+public class ActionUtil {
+
+	public static void getUserGroup(HttpServletRequest request)
+		throws Exception {
+
+		long userGroupId = ParamUtil.getLong(request, "userGroupId");
+
+		UserGroup userGroup = null;
+
+		if (userGroupId > 0) {
+			userGroup = UserGroupServiceUtil.getUserGroup(userGroupId);
+		}
+
+		request.setAttribute(WebKeys.USER_GROUP, userGroup);
+	}
+
+	public static void getUserGroup(PortletRequest portletRequest)
+		throws Exception {
+
+		HttpServletRequest request = PortalUtil.getHttpServletRequest(
+			portletRequest);
+
+		getUserGroup(request);
+	}
+}
