+++ portal-impl/src/com/liferay/portlet/usergroupsadmin/action/EditUserGroupAction.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.usersadmin.action;
+package com.liferay.portlet.usergroupsadmin.action;
 
 import com.liferay.portal.DuplicateUserGroupException;
 import com.liferay.portal.NoSuchUserGroupException;
@@ -65,7 +65,7 @@ public class EditUserGroupAction extends PortletAction {
 			if (e instanceof PrincipalException) {
 				SessionErrors.add(actionRequest, e.getClass().getName());
 
-				setForward(actionRequest, "portlet.users_admin.error");
+				setForward(actionRequest, "portlet.user_groups_admin.error");
 			}
 			else if (e instanceof DuplicateUserGroupException ||
 					 e instanceof NoSuchUserGroupException ||
@@ -104,7 +104,7 @@ public class EditUserGroupAction extends PortletAction {
 
 				SessionErrors.add(renderRequest, e.getClass().getName());
 
-				return mapping.findForward("portlet.users_admin.error");
+				return mapping.findForward("portlet.user_groups_admin.error");
 			}
 			else {
 				throw e;
@@ -112,7 +112,7 @@ public class EditUserGroupAction extends PortletAction {
 		}
 
 		return mapping.findForward(getForward(
-			renderRequest, "portlet.users_admin.edit_user_group"));
+			renderRequest, "portlet.user_groups_admin.edit_user_group"));
 	}
 
 	protected void deleteUserGroups(ActionRequest actionRequest)
