+++ portal-impl/src/content/Language.properties
@@ -9,6 +9,8 @@ lang.line.end=right
 ##
 ## Portlet descriptions and titles
 ##
+javax.portlet.description.127=User groups provide a way to group users independently of the organizations to which they belong. Administrators can define a user group and assign the user group as a member of a site to make all of its users members automatically.
+javax.portlet.title.127=User Groups
 
 javax.portlet.description.104=Update Manager shows you all user installed plugins as well as their version number, whether it is from the Liferay repository, and whether it can be upgraded.
 javax.portlet.description.125=Users are individuals who perform tasks using the portal. Administrators can create new users or deactivate existing users. Users can join sites, be placed into organization hierarchies, be delegated permissions in roles, or be collected into user groups. User groups can cross organizational boundaries, and can be used to assign all members to other collections, such as sites or roles. User groups may also be used to customize personal site templates for members.
