+++ portal-service/src/com/liferay/portal/util/PortletKeys.java
@@ -218,6 +218,8 @@ public class PortletKeys {
 
 	public static final String UPDATE_MANAGER = "104";
 
+	public static final String USER_GROUPS_ADMIN = "127";
+
 	public static final String USERS_ADMIN = "125";
 
 	public static final String WIKI = "36";
