+++ portal-web/docroot/WEB-INF/liferay-display.xml
@@ -53,6 +53,7 @@
 		<portlet id="104" />
 		<portlet id="111" />
 		<portlet id="125" />
+        <portlet id="127" />
 		<portlet id="128" />
 		<portlet id="129" />
 		<portlet id="130" />
