+++ portal-web/docroot/WEB-INF/portlet-custom.xml
@@ -1734,6 +1734,23 @@
 		</security-role-ref>
 	</portlet>
 	<portlet>
+		<portlet-name>127</portlet-name>
+		<display-name>User Groups Admin</display-name>
+		<portlet-class>com.liferay.portlet.StrutsPortlet</portlet-class>
+		<init-param>
+			<name>view-action</name>
+			<value>/user_groups_admin/view</value>
+		</init-param>
+		<expiration-cache>0</expiration-cache>
+		<supports>
+			<mime-type>text/html</mime-type>
+		</supports>
+		<resource-bundle>com.liferay.portlet.StrutsResourceBundle</resource-bundle>
+		<security-role-ref>
+			<role-name>administrator</role-name>
+		</security-role-ref>
+	</portlet>
+	<portlet>
 		<portlet-name>128</portlet-name>
 		<display-name>Roles Admin</display-name>
 		<portlet-class>com.liferay.portlet.StrutsPortlet</portlet-class>
