+++ portal-web/docroot/html/portlet/user_groups_admin/select_user_group.jsp
@@ -14,14 +14,14 @@
  */
 --%>
 
-<%@ include file="/html/portlet/users_admin/init.jsp" %>
+<%@ include file="/html/portlet/user_groups_admin/init.jsp" %>
 
 <%
 String target = ParamUtil.getString(request, "target");
 
 PortletURL portletURL = renderResponse.createRenderURL();
 
-portletURL.setParameter("struts_action", "/users_admin/select_user_group");
+portletURL.setParameter("struts_action", "/user_groups_admin/select_user_group");
 %>
 
 <aui:form action="<%= portletURL.toString() %>" method="post" name="fm">
@@ -33,7 +33,7 @@ portletURL.setParameter("struts_action", "/users_admin/select_user_group");
 		searchContainer="<%= new UserGroupSearch(renderRequest, portletURL) %>"
 	>
 		<liferay-ui:search-form
-			page="/html/portlet/users_admin/user_group_search.jsp"
+			page="/html/portlet/user_groups_admin/user_group_search.jsp"
 		/>
 
 		<%
@@ -46,7 +46,8 @@ portletURL.setParameter("struts_action", "/users_admin/select_user_group");
 			if (filterManageableUserGroups) {
 				List<UserGroup> userGroups = UserGroupLocalServiceUtil.search(company.getCompanyId(), searchTerms.getName(), searchTerms.getDescription(), null, QueryUtil.ALL_POS, QueryUtil.ALL_POS, searchContainer.getOrderByComparator());
 
-				userGroups = UsersAdminUtil.filterUserGroups(permissionChecker, userGroups);
+				userGroups = UsersAdminUtil.filterUserGroups(
+						permissionChecker, userGroups);
 
 				total = userGroups.size();
 				results = ListUtil.subList(userGroups, searchContainer.getStart(), searchContainer.getEnd());
