+++ portal-web/docroot/html/portlet/user_groups_admin/user_group_search.jsp
@@ -1,3 +1,4 @@
+<%--
 /**
  * Copyright (c) 2000-2011 Liferay, Inc. All rights reserved.
  *
@@ -11,23 +12,18 @@
  * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
  * details.
  */
+--%>
 
-package com.liferay.portlet.usersadmin.search;
+<%@ include file="/html/portlet/user_groups_admin/init.jsp" %>
 
-import com.liferay.portal.kernel.dao.search.DAOParamUtil;
+<%
+UserGroupSearch searchContainer = (UserGroupSearch)request.getAttribute("liferay-ui:search:searchContainer");
 
-import javax.portlet.PortletRequest;
+UserGroupDisplayTerms displayTerms = (UserGroupDisplayTerms)searchContainer.getDisplayTerms();
+%>
 
-/**
- * @author Charles May
- */
-public class UserGroupSearchTerms extends UserGroupDisplayTerms {
+<span class="aui-search-bar lfr-display-terms-search">
+	<aui:input inlineField="<%= true %>" label="" name="<%= displayTerms.NAME %>" size="30" type="text" value="<%= displayTerms.getName() %>" />
 
-	public UserGroupSearchTerms(PortletRequest portletRequest) {
-		super(portletRequest);
-
-		description = DAOParamUtil.getLike(portletRequest, DESCRIPTION);
-		name = DAOParamUtil.getLike(portletRequest, NAME);
-	}
-
-}
+	<aui:button type="submit" value="search" />
+</span>
