+++ portal-web/docroot/html/portlet/users_admin/toolbar.jsp
@@ -23,13 +23,15 @@ String toolbarItem = ParamUtil.getString(request, "toolbarItem", "view");
 <div class="lfr-portlet-toolbar">
 	<portlet:renderURL var="viewUsersURL">
 		<portlet:param name="struts_action" value="/users_admin/view" />
+		<portlet:param name="usersListView" value="<%= UserConstants.LIST_VIEW_TREE %>" />
+		<portlet:param name="saveUsersListView" value="<%= Boolean.TRUE.toString() %>" />
 	</portlet:renderURL>
 
-	<liferay-ui:icon-menu align="left" cssClass='<%= "lfr-toolbar-button add-button " + (toolbarItem.equals("view") ? "current" : StringPool.BLANK) %>' direction="down" extended="<%= false %>" icon='<%= themeDisplay.getPathThemeImages() + "/common/view_users.png" %>' message="view">
-		<%@ include file="/html/portlet/users_admin/user/list_views.jspf" %>
-	</liferay-ui:icon-menu>
+	<span class="lfr-toolbar-button view-button <%= toolbarItem.equals("view-all") ? "current" : StringPool.BLANK %>">
+		<a href="<%= viewUsersURL %>"><liferay-ui:message key="view-all" /></a>
+	</span>
 
-	<c:if test="<%= PortalPermissionUtil.contains(permissionChecker, ActionKeys.ADD_ORGANIZATION) || PortalPermissionUtil.contains(permissionChecker, ActionKeys.ADD_USER) || PortalPermissionUtil.contains(permissionChecker, ActionKeys.ADD_USER_GROUP) %>">
+	<c:if test="<%= PortalPermissionUtil.contains(permissionChecker, ActionKeys.ADD_ORGANIZATION) || PortalPermissionUtil.contains(permissionChecker, ActionKeys.ADD_USER) %>">
 		<liferay-ui:icon-menu align="left" cssClass='<%= "lfr-toolbar-button add-button " + (toolbarItem.equals("add") ? "current" : StringPool.BLANK) %>' direction="down" extended="<%= false %>" icon='<%= themeDisplay.getPathThemeImages() + "/common/add.png" %>' message="add">
 			<c:if test="<%= PortalPermissionUtil.contains(permissionChecker, ActionKeys.ADD_USER) %>">
 				<portlet:renderURL var="addUserURL">
@@ -66,19 +68,6 @@ String toolbarItem = ParamUtil.getString(request, "toolbarItem", "view");
 				}
 				%>
 
-			</c:if>
-
-			<c:if test="<%= PortalPermissionUtil.contains(permissionChecker, ActionKeys.ADD_USER_GROUP) %>">
-				<portlet:renderURL var="addUsergroupURL">
-					<portlet:param name="struts_action" value="/users_admin/edit_user_group" />
-					<portlet:param name="redirect" value="<%= currentURL %>" />
-				</portlet:renderURL>
-
-				<liferay-ui:icon
-					image="assign_user_group_roles"
-					message="user-group"
-					url="<%= addUsergroupURL %>"
-				/>
 			</c:if>
 		</liferay-ui:icon-menu>
 	</c:if>
