--- portal-web/docroot/html/portlet/users_admin/user/list_views.jspf
@@ -1,49 +1,0 @@
-<%--
-/**
- * Copyright (c) 2000-2011 Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
---%>
-
-<c:if test="<%= PropsValues.USERS_LIST_VIEWS.length > 1 %>">
-	<c:if test="<%= ArrayUtil.contains(PropsValues.USERS_LIST_VIEWS, UserConstants.LIST_VIEW_TREE) %>">
-		<portlet:renderURL var="viewUsersURL">
-			<portlet:param name="struts_action" value="/users_admin/view" />
-			<portlet:param name="tabs1" value="users" />
-			<portlet:param name="usersListView" value="<%= UserConstants.LIST_VIEW_TREE %>" />
-			<portlet:param name="saveUsersListView" value="<%= Boolean.TRUE.toString() %>" />
-		</portlet:renderURL>
-
-		<liferay-ui:icon
-			image="view_users"
-			message="users-and-organizations"
-			method="get"
-			url="<%= viewUsersURL %>"
-		/>
-	</c:if>
-
-	<c:if test="<%= ArrayUtil.contains(PropsValues.USERS_LIST_VIEWS, UserConstants.LIST_VIEW_FLAT_USER_GROUPS) %>">
-		<portlet:renderURL var="viewUsergroupsURL">
-			<portlet:param name="struts_action" value="/users_admin/view" />
-			<portlet:param name="tabs1" value="users" />
-			<portlet:param name="usersListView" value="<%= UserConstants.LIST_VIEW_FLAT_USER_GROUPS%>" />
-			<portlet:param name="saveUsersListView" value="<%= Boolean.TRUE.toString() %>" />
-		</portlet:renderURL>
-
-		<liferay-ui:icon
-			image="assign_user_group_roles"
-			message="user-groups"
-			method="get"
-			url="<%= viewUsergroupsURL %>"
-		/>
-	</c:if>
-</c:if>
